Rename module "Utility" to "utils"
Following PEP8: https://peps.python.org/pep-0008/#package-and-module-names
