+++ src/sdx/pce/load_balancing/TE_Solver.py
@@ -13,8 +13,8 @@ import networkx as nx
 import numpy as np
 from ortools.linear_solver import pywraplp
 
-import sdx.pce.Utility.global_name as global_name
-from sdx.pce.Utility.functions import GraphFunction
+import sdx.pce.utils.global_name as global_name
+from sdx.pce.utils.functions import GraphFunction
 
 
 class TE_Solver:
