+++ src/sdx/pce/utils/randomConnectionGenerator.py
