+++ tests/test_te_solver.py
@@ -3,10 +3,10 @@ import unittest
 
 import networkx as nx
 
-import sdx.pce.Utility.global_name as global_name
+import sdx.pce.utils.global_name as global_name
 from sdx.pce.load_balancing.TE_Solver import TE_Solver
-from sdx.pce.Utility.randomConnectionGenerator import RandomConnectionGenerator
-from sdx.pce.Utility.randomTopologyGenerator import RandomTopologyGenerator, dot_file
+from sdx.pce.utils.randomConnectionGenerator import RandomConnectionGenerator
+from sdx.pce.utils.randomTopologyGenerator import RandomTopologyGenerator, dot_file
 
 Connection = "./tests/data/test_connection.json"
 Solution = "./tests/data/test_MC_solution.json"
